requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.2.1
Django==4.1.7
django-filter==23.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
pyrsistent==0.19.3
pytest==7.2.2
pytest-django==4.5.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
Unidecode==1.3.6
uritemplate==4.1.1
